+++ core/src/main/kotlin/org/jetbrains/research/testspark/core/test/TestCompiler.kt
@@ -1,7 +1,6 @@
 package org.jetbrains.research.testspark.core.test
 
 import org.jetbrains.research.testspark.core.test.data.TestCaseGeneratedByLLM
-import org.jetbrains.research.testspark.core.test.strategies.TestCompilerStrategy
 import org.jetbrains.research.testspark.core.utils.DataFilesUtil
 
 data class TestCasesCompilationResult(
@@ -9,11 +8,9 @@ data class TestCasesCompilationResult(
     val compilableTestCases: MutableSet<TestCaseGeneratedByLLM>,
 )
 
-class TestCompiler(
-    private val javaHomeDirectoryPath: String,
+abstract class TestCompiler(
     private val libPaths: List<String>,
     private val junitLibPaths: List<String>,
-    private val language: Language,
 ) {
     fun compileTestCases(
         generatedTestCasesPaths: List<String>,
@@ -42,20 +39,7 @@ class TestCompiler(
      * @return A pair containing a boolean value indicating whether the compilation was successful (true) or not (false),
      *         and a string message describing any error encountered during compilation.
      */
-    fun compileCode(path: String, projectBuildPath: String): Pair<Boolean, String> {
-        return when (language) {
-            Language.Java -> TestCompilerStrategy.compileJavaCode(
-                path,
-                "\"${getClassPaths(projectBuildPath)}\"",
-                javaHomeDirectoryPath,
-            )
-
-            Language.Kotlin -> TestCompilerStrategy.compileKotlinCode(
-                path,
-                "\"${getClassPaths(projectBuildPath)}\"",
-            )
-        }
-    }
+    abstract fun compileCode(path: String, projectBuildPath: String): Pair<Boolean, String>
 
     /**
      * Generates the path for the command by concatenating the necessary paths.
