+++ core/src/main/kotlin/org/jetbrains/research/testspark/core/test/data/TestCaseGeneratedByLLM.kt
@@ -1,5 +1,7 @@
 package org.jetbrains.research.testspark.core.test.data
 
+import org.jetbrains.research.testspark.core.test.PrintTestBody
+
 /**
  *
  * Represents a test case generated by LLM.
@@ -11,7 +13,7 @@ data class TestCaseGeneratedByLLM(
     var expectedException: String = "",
     var throwsException: String = "",
     var lines: MutableList<TestLine> = mutableListOf(),
-    val printTestBodyStrategy: PrintTestBodyStrategy,
+    val printTestBodyStrategy: PrintTestBody,
 ) {
 
     /**
