+++ core/src/main/kotlin/org/jetbrains/research/testspark/core/test/kotlin/KotlinJUnitTestSuiteParser.kt
@@ -2,8 +2,8 @@ package org.jetbrains.research.testspark.core.test.kotlin
 
 import org.jetbrains.research.testspark.core.data.JUnitVersion
 import org.jetbrains.research.testspark.core.test.TestSuiteParser
-import org.jetbrains.research.testspark.core.test.TestSuiteParserStrategy
 import org.jetbrains.research.testspark.core.test.data.TestSuiteGeneratedByLLM
+import org.jetbrains.research.testspark.core.test.strategies.TestSuiteParserStrategy
 
 class KotlinJUnitTestSuiteParser(
     private val packageName: String,
