+++ core/src/main/kotlin/org/jetbrains/research/testspark/core/test/strategies/JUnitTestSuiteParserStrategy.kt
@@ -1,9 +1,8 @@
 package org.jetbrains.research.testspark.core.test.strategies
 
 import org.jetbrains.research.testspark.core.data.JUnitVersion
+import org.jetbrains.research.testspark.core.test.PrintTestBody
 import org.jetbrains.research.testspark.core.test.TestCaseParseResult
-import org.jetbrains.research.testspark.core.test.TestSuiteParserStrategy
-import org.jetbrains.research.testspark.core.test.data.PrintTestBodyStrategy
 import org.jetbrains.research.testspark.core.test.data.TestCaseGeneratedByLLM
 import org.jetbrains.research.testspark.core.test.data.TestLine
 import org.jetbrains.research.testspark.core.test.data.TestLineType
@@ -16,7 +15,7 @@ class JUnitTestSuiteParserStrategy : TestSuiteParserStrategy {
         importPattern: Regex,
         packageName: String,
         testNamePattern: String,
-        printTestBodyStrategy: PrintTestBodyStrategy,
+        printTestBodyStrategy: PrintTestBody,
     ): TestSuiteGeneratedByLLM? {
         if (rawText.isBlank()) {
             return null
@@ -87,7 +86,7 @@ private class JUnitTestCaseParser {
         rawTest: String,
         isLastTestCaseInTestSuite: Boolean,
         testNamePattern: String,
-        printTestBodyStrategy: PrintTestBodyStrategy,
+        printTestBodyStrategy: PrintTestBody,
     ): TestCaseParseResult {
         var expectedException = ""
         var throwsException = ""
