--- kotlin/src/main/kotlin/org/jetbrains/research/testspark/kotlin/KotlinBuilderHelper.kt
@@ -1,6 +1,0 @@
-package org.jetbrains.research.testspark.kotlin
-
-class KotlinBuilderHelper {
-    companion object {
-    }
-}
